update legal contract change column relate

+++ app/Http/Controllers/Legal/ContractRequest/MouldController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers\Legal\ContractRequest;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Legal\StoreMould;
+use App\Models\Legal\LegalComercialTerm;
+use App\Models\Legal\LegalContractDest;
 use App\Services\Legal\Interfaces\ComercialListsServiceInterface;
 use App\Services\Legal\Interfaces\ComercialTermServiceInterface;
 use App\Services\Legal\Interfaces\ContractDescServiceInterface;
+use App\Services\Legal\Interfaces\ContractRequestServiceInterface;
 use App\Services\Legal\Interfaces\PaymentTypeServiceInterface;
 use App\Services\Utils\FileService;
 use Illuminate\Http\Request;
@@ -19,19 +22,21 @@ class MouldController extends Controller
     protected $paymentTypeService;
     protected $fileService;
     protected $comercialListsService;
-    protected $comercialTermService;
+    protected $comercialTermService, $contractRequestService;
     public function __construct(
         ContractDescServiceInterface $contractDescServiceInterface,
         PaymentTypeServiceInterface $paymentTypeServiceInterface,
         FileService $fileService,
         ComercialListsServiceInterface $comercialListsServiceInterface,
-        ComercialTermServiceInterface $comercialTermServiceInterface
+        ComercialTermServiceInterface $comercialTermServiceInterface,
+        ContractRequestServiceInterface $contractRequestService
     ) {
         $this->contractDescService = $contractDescServiceInterface;
         $this->paymentTypeService = $paymentTypeServiceInterface;
         $this->fileService = $fileService;
         $this->comercialListsService = $comercialListsServiceInterface;
         $this->comercialTermService = $comercialTermServiceInterface;
+        $this->contractRequestService = $contractRequestService;
     }
     /**
      * Display a listing of the resource.
@@ -59,9 +64,25 @@ class MouldController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreMould $request)
     {
-        return \redirect()->route('legal.contract-request.mould.index');
+        $dest = $request->only('quotation', 'coparation_sheet', 'purchase_order', 'drawing', 'contract_id', 'value_of_contract', 'payment_type_id', 'warranty');
+        // comercialTerm data
+        $term = $request->only('scope_of_work', 'to_manufacture', 'of', 'purchase_order_no', 'quotation_no', 'dated', 'delivery_date');
+
+        DB::beginTransaction();
+        try {
+            $contract_desc = new LegalContractDest($dest);
+            $contract_desc->save();
+            $term['contract_dest_id'] = $contract_desc->id;
+            $contract_term = new LegalComercialTerm($term);
+            $contract_term->save();
+            DB::commit();
+            return \redirect()->route('legal.contract-request.show', $contract_desc->contract_id);
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return \redirect()->back()->with('error', "Error : " . $e->getMessage());
+        }
     }
 
     /**
@@ -84,16 +105,24 @@ class MouldController extends Controller
     public function edit($id)
     {
         try {
-            $mould = $this->contractDescService->search($id);
+            // $mould = $this->contractDescService->search($id);
 
-            if ($mould->value_of_contract) {
-                $mould->value_of_contract = explode(",", $mould->value_of_contract);
+            // if ($mould->value_of_contract) {
+            //     $mould->value_of_contract = explode(",", $mould->value_of_contract);
+            // }
+            // $paymentType = $this->paymentTypeService->dropdown($mould->legalcontract->agreement_id);
+
+            $contract = $this->contractRequestService->find($id);
+            $paymentType = $this->paymentTypeService->dropdown($contract->agreement_id);
+            if ($contract->legalContractDest) {
+                $contract->legalContractDest->value_of_contract = explode(",", $contract->legalContractDest->value_of_contract);
+                return \view('legal.ContractRequestForm.Mould.edit')->with(['contract' => $contract, 'paymentType' => $paymentType]);
+            } else {
+                return \view('legal.ContractRequestForm.Mould.create', \compact('contract', 'paymentType'));
             }
-            $paymentType = $this->paymentTypeService->dropdown($mould->legalcontract->agreement_id);
         } catch (\Exception $e) {
             return \redirect()->back()->with('error', "Error : " . $e->getMessage());
         }
-        return \view('legal.ContractRequestForm.Mould.edit')->with(['mould' => $mould, 'paymentType' => $paymentType]);
     }
 
     /**
@@ -105,38 +134,21 @@ class MouldController extends Controller
      */
     public function update(StoreMould $request, $id)
     {
-        $data = $request->except(['_token', '_method']);
-        $attributes = [];
-        $comercialAttr = [];
-
-        $attributes['purchase_order'] = $data['purchase_order'];
-        $attributes['quotation'] = $data['quotation'];
-        $attributes['coparation_sheet'] = $data['coparation_sheet'];
-        $attributes['drawing'] = $data['drawing'];
-
-        $attributes['payment_type_id'] = (int) $data['payment_type_id'];
-        $attributes['value_of_contract'] = $data['value_of_contract'];
-        $attributes['warranty'] = (int) $data['warranty'];
+        $dest = $request->only('quotation', 'coparation_sheet', 'purchase_order', 'drawing', 'contract_id', 'value_of_contract', 'payment_type_id', 'warranty');
         // comercialTerm data
-        $comercialAttr['scope_of_work'] = $data['scope_of_work'];
-        $comercialAttr['to_manufacture'] = $data['to_manufacture'];
-        $comercialAttr['of'] = $data['of'];
-        $comercialAttr['purchase_order_no'] = $data['purchase_order_no'];
-        $comercialAttr['quotation_no'] = $data['quotation_no'];
-        $comercialAttr['dated'] = $data['dated'];
-        $comercialAttr['delivery_date'] = $data['delivery_date'];
+        $term = $request->only('scope_of_work', 'to_manufacture', 'of', 'purchase_order_no', 'quotation_no', 'dated', 'delivery_date');
 
         DB::beginTransaction();
         try {
-            if ($data['comercial_term_id']) {
-                $this->comercialTermService->update($comercialAttr, $data['comercial_term_id']);
-                $attributes['comercial_term_id'] = $data['comercial_term_id'];
-            } else {
-                $attributes['comercial_term_id'] = $this->comercialTermService->create($comercialAttr)->id;
-            }
             $mould = $this->contractDescService->find($id);
-            $this->contractDescService->update($attributes, $id);
-            // \dd($attributes['drawing'],$mould->drawing,$this->contractDescService->find($id)->drawing);
+            if ($mould->legalContract->legalComercialList->count() < 1) {
+                return \redirect()->back()->with('error', "Error : ");
+            }
+            if ($mould->legalComercialTerm) {
+                $this->comercialTermService->update($term, $mould->legalComercialTerm->id);
+            }
+
+            $this->contractDescService->update($dest, $mould->id);
             if ($mould->purchase_order !== $request->purchase_order) {
                 Storage::delete($mould->purchase_order);
             }
@@ -155,7 +167,7 @@ class MouldController extends Controller
             return \redirect()->back()->with('error', "Error : " . $e->getMessage());
         }
         DB::commit();
-        return \redirect()->route('legal.contract-request.index');
+        return \redirect()->route('legal.contract-request.show', $mould->contract_id);
     }
 
     /**
